requirements.txt
openai-whisper
torch
numpy
sounddevice
scipy
gtts
